Use gimme to get known versions list
and cut down on intermediate and platform-specific lists
