+++ .gitignore
@@ -1,2 +1,2 @@
-/.testdata/object-urls
+/.testdata/known-versions.txt
 /gimme-generate
