--- .testdata/sample-binary-linux
@@ -1,17 +1,0 @@
-# WARNING: file generated by gnomes DO NOT EDIT
-#
-# These are versions which are passed to runtests, on each architecture, for CI
-# testing. These do not need to be authoritative.
-#
-# Listing all versions ever provides for quadratic test runtime as we expand
-# architectures and OSes. So we no longer do that.
-#
-# Instead we include all versions of the current release series plus two.
-
-#
-
-1.19.2
-1.18.7
-1.17.13
-1.16.15
-1.15.15
